+++ src/Cache/File.php
@@ -48,6 +48,7 @@ namespace SimplePie\Cache;
  *
  * @package SimplePie
  * @subpackage Caching
+ * @deprecated since SimplePie 1.8, use implementation of "Psr\SimpleCache\CacheInterface" instead
  */
 class File implements Base
 {